# [Dashboard] Speed up clusters page first load by 20x for long cluster history