@Michaelvll 2025-09-16
Nice! We should make the cluster history paginated as well